requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml>=5.2.0
python-dotenv==1.0.1
pandas>=2.2.3
streamlit>=1.35.0
plotly>=5.22.0
openpyxl>=3.1.0
APScheduler>=3.10.0
pytest==8.1.1
pytest-cov==5.0.0
